CR : fix cast to prevent future warning

+++ libs/dimg/filters/auto/stretchfilter.cpp
@@ -6,7 +6,7 @@
  * Date        : 2005-24-01
  * Description : stretch contrast image filter.
  *
- * Copyright (C) 2005-2012 by Gilles Caulier <caulier dot gilles at gmail dot com>
+ * Copyright (C) 2005-2013 by Gilles Caulier <caulier dot gilles at gmail dot com>
  *
  * This program is free software; you can redistribute it
  * and/or modify it under the terms of the GNU General
@@ -106,7 +106,7 @@ void StretchFilter::stretchContrastImage()
 
     // Find the histogram boundaries by locating the 0.1 percent levels.
 
-    number_pixels       = (long long)(m_refImage.width() * m_refImage.height());
+    number_pixels       = (long long)m_refImage.width() * (long long)m_refImage.height();
     threshold_intensity = number_pixels / 1000;
 
     memset(&high, 0, sizeof(struct double_packet));
